+++ HISTORY.md
@@ -50,6 +50,7 @@
 ### 2025-12-16
 * objects: fixed a costumePoint() glitch
 * objects: automatically grow a target costume when "overdrawing" it
+* objects: enable "overdrawing" a costume-less sprite
 
 ### 2025-12-14
 * gui: fixed a zooming layout glitch